wip: remove @islands/frontmatter, move functionality to @islands/mdx
