--- packages/frontmatter/CHANGELOG.md
@@ -1,78 +1,0 @@
-## [0.4.2](https://github.com/ElMassimo/iles/compare/frontmatter@0.4.1...frontmatter@0.4.2) (2021-12-06)
-
-
-### Features
-
-* add `prettyUrls: false` to support services without transparent redirects ([3185aea](https://github.com/ElMassimo/iles/commit/3185aeaa6d7c4e49ec3da0ae60252ab660a70c6c))
-
-
-
-## [0.4.1](https://github.com/ElMassimo/iles/compare/frontmatter@0.4.0...frontmatter@0.4.1) (2021-11-30)
-
-
-### Features
-
-* add withImageSrc option to `markdown` to enable imagetools customization ([6da38bb](https://github.com/ElMassimo/iles/commit/6da38bbe218f53505cd6acb04563e6342b67c66a))
-
-
-
-# [0.4.0](https://github.com/ElMassimo/iles/compare/frontmatter@0.3.0...frontmatter@0.4.0) (2021-11-28)
-
-
-### Bug Fixes
-
-* explicitly catch any import errors that may occur ([821d6cf](https://github.com/ElMassimo/iles/commit/821d6cf4b93d0c676fd29c0b627deca5a697a241))
-
-
-### Features
-
-* import images in MDX files to enable asset fingerprinting ([832855b](https://github.com/ElMassimo/iles/commit/832855b4a19ac67b572074ba7613cc46e7a6c552))
-
-
-
-# [0.3.0](https://github.com/ElMassimo/iles/compare/frontmatter@0.2.0...frontmatter@0.3.0) (2021-11-02)
-
-
-
-# [0.2.0](https://github.com/ElMassimo/iles/compare/frontmatter@0.0.3-0...frontmatter@0.2.0) (2021-11-02)
-
-- __BREAKING CHANGES__: Now it's an îles plugin that registers the remark plugin.
-
-## 0.0.3-0 (2021-10-27)
-
-
-### Bug Fixes
-
-* drop remark-frontmatter and add micromark plugins directly ([8d6f3f3](https://github.com/ElMassimo/iles/commit/8d6f3f3b184674e30181a7ca52361de3baaeb5ac))
-
-
-### Features
-
-* Docs + Layout Shortcuts + Meta + Site + SEO Tags ([#3](https://github.com/ElMassimo/iles/issues/3)) ([4ed90b7](https://github.com/ElMassimo/iles/commit/4ed90b72cf354823f023dd09f6797b8b71cff35b))
-* extendFrontmatter hook. unified usage as route.meta.frontmatter ([31062e0](https://github.com/ElMassimo/iles/commit/31062e04193822cddf1ef9069bec9c448b6d3b72))
-* provide filename in extendFrontmatter hook ([35b55c6](https://github.com/ElMassimo/iles/commit/35b55c6505561c0151960697a7bcebc4953ad968))
-* use remark plugin to accurately wrap islands in MDX ([946240d](https://github.com/ElMassimo/iles/commit/946240d9ab0e90536b0aa6e3d6d3b4e564cecc07))
-
-
-
-## 0.0.2 (2021-10-08)
-
-
-### Features
-
-* Docs + Layout Shortcuts + Meta + Site + SEO Tags ([#3](https://github.com/ElMassimo/iles/issues/3)) ([4ed90b7](https://github.com/ElMassimo/iles/commit/4ed90b72cf354823f023dd09f6797b8b71cff35b))
-* extendFrontmatter hook. unified usage as route.meta.frontmatter ([31062e0](https://github.com/ElMassimo/iles/commit/31062e04193822cddf1ef9069bec9c448b6d3b72))
-* provide filename in extendFrontmatter hook ([35b55c6](https://github.com/ElMassimo/iles/commit/35b55c6505561c0151960697a7bcebc4953ad968))
-* use remark plugin to accurately wrap islands in MDX ([946240d](https://github.com/ElMassimo/iles/commit/946240d9ab0e90536b0aa6e3d6d3b4e564cecc07))
-
-
-
-## 0.0.1 (2021-09-16)
-
-
-### Features
-
-* extendFrontmatter hook. unified usage as route.meta.frontmatter ([31062e0](https://github.com/maximomussini/iles/commit/31062e04193822cddf1ef9069bec9c448b6d3b72))
-
-
-
