--- packages/frontmatter/README.md
@@ -1,33 +1,0 @@
-<p align="center">
-  <a href="https://iles-docs.netlify.app">
-    <img src="https://github.com/ElMassimo/iles/blob/main/docs/images/banner.png"/>
-  </a>
-</p>
-
-<p align="center">
-<table>
-<tbody>
-<td align="center">
-<br/>
-<p align="center">
-  <h3><samp>@islands/frontmatter</samp></h3>
-  <img width="2000" height="0">
-</p>
-</td>
-</tbody>
-</table>
-</p>
-
-[îles]: https://github.com/ElMassimo/iles
-[docs]: https://iles-docs.netlify.app
-[remark]: https://github.com/remarkjs/remark
-[frontmatter]: https://iles-docs.netlify.app/guide/markdown#frontmatter-and-meta
-[mdx documents]: https://iles-docs.netlify.app/guide/markdown
-
-An [îles] module that injects a [remark] plugin to parse [frontmatter] in [MDX documents]:
-
-- ⚙️ runs `extendFrontmatter` to add additional `frontmatter` or `meta` info
-
-- 🧱 makes each frontmatter property available for bindings
-
-- 📃 sets a `frontmatter` object with the full content
