--- packages/frontmatter/package.json
@@ -1,45 +1,0 @@
-{
-  "name": "@islands/frontmatter",
-  "version": "0.4.2",
-  "scripts": {
-    "dev": "npm run build -- --watch",
-    "build": "tsup src/frontmatter.ts",
-    "prepublishOnly": "npm run build"
-  },
-  "type": "module",
-  "files": [
-    "dist",
-    "src"
-  ],
-  "types": "dist/frontmatter.d.ts",
-  "main": "dist/frontmatter.js",
-  "exports": {
-    ".": {
-      "import": "./dist/frontmatter.js",
-      "require": "./src/frontmatter.cjs"
-    },
-    "./package.json": "./package.json"
-  },
-  "funding": "https://github.com/sponsors/ElMassimo",
-  "author": "Máximo Mussini <maximomussini@gmail.com>",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/ElMassimo/iles"
-  },
-  "homepage": "https://github.com/ElMassimo/iles",
-  "bugs": "https://github.com/ElMassimo/iles/issues",
-  "dependencies": {
-    "estree-util-is-identifier-name": "^2.0.0",
-    "estree-util-value-to-estree": "^1.3.0",
-    "js-yaml": "^4.1",
-    "mdast-util-frontmatter": "^1.0.0",
-    "micromark-extension-frontmatter": "^1.0.0"
-  },
-  "devDependencies": {
-    "@types/estree": "^0.0.46",
-    "@types/js-yaml": "^4.0.0",
-    "tsup": "5.5.0",
-    "typescript": "^4.4.4",
-    "unified": "^10.1.0"
-  }
-}
