--- packages/frontmatter/src/frontmatter.cjs
@@ -1,5 +1,0 @@
-module.exports = (...args) => new Promise((resolve, reject) => {
-  import('../dist/frontmatter.js')
-    .then(m => resolve(m.default(...args)))
-    .catch(reject)
-})
