--- packages/frontmatter/tsup.config.ts
@@ -1,7 +1,0 @@
-import type { Options } from 'tsup'
-export const tsup: Options = {
-  dts: true,
-  target: 'node14',
-  splitting: true,
-  format: ['esm'],
-}
